Added aurelia-logging to source under thirdparty, temp workaround to keep builds functioning while figuring out how best to handle EC_Logger in this project.
Moved aurelia-logging definitions to the thirdparty folder.

Added new build.sh to build the typescript, and copy third party dependencies to the dist/declarations.

soft disabled the join descriptor for the record while researching the subrecord descriptor to see what would be required to update that for read/write support and possible alternative aliasing for instances where you might not want to pull the record every time.

Record.ts dropped sublistline import and line constructor export. This was only used in the sublist descriptor which has already been moved out.

Sublist.ts removed the NonFunctionPropertyName definition. This was not referenced in the project.

Added custom record example.